requirements.txt
adal==1.2.7
aniso8601==9.0.1
attrs==24.3.0
blinker==1.9.0
boto3==1.35.39
botocore==1.35.39
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
circuitbreaker==2.0.0
click==8.1.7
configparser==7.1.0
cryptography==43.0.3
debtcollector==3.0.0
dnspython==2.7.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
flasgger==0.9.7.1
Flask==3.0.3
Flask-RESTful==0.3.9
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.68.1
grpcio-status==1.68.1
httplib2==0.22.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mistune==3.0.2
mock==5.1.0
msgpack==1.1.0
netaddr==1.3.0
numpy==2.0.2
oauth2client==4.1.3
oci==2.136.0
openpyxl==3.2.0b1
oslo.config==9.6.0
oslo.context==5.7.0
oslo.i18n==6.5.0
oslo.log==6.1.2
oslo.serialization==5.6.0
oslo.utils==8.0.0
packaging==24.2
pandas==2.2.3
pbr==6.1.0
pluggy==1.5.0
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pyinstrument==4.6.2
PyJWT==2.9.0
pymongo==4.10.1
pyOpenSSL==24.3.0
pyotp==2.9.0
pyparsing==3.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rfc3339==6.2
rfc3986==2.0.0
rpds-py==0.22.3
rsa==4.9
s3transfer==0.10.4
six==1.16.0
stevedore==5.4.0
tomli==2.2.1
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.20
Werkzeug==3.0.4
wrapt==1.17.0
xmltodict==0.14.1
zipp==3.21.0
